refactor(settings_ui): lift fs-custom-paths handlers to _handlers_fs.py
Establishes the module-level-handler + thin-factory pattern: nested
closures make the factory's C901 sum all handler complexity, so handlers
are module-level (own budget) and build_fs_handlers binds server into
request-only wrappers.
