+++ src/ha_mcp/settings_ui/__init__.py
@@ -39,6 +39,7 @@ from ..llm_exposure import LLM_API_CONFIG_KEY
 from ..transforms import DEFAULT_PINNED_TOOLS
 from ..utils.data_paths import get_data_dir
 from . import _persistence, _supervisor
+from ._handlers_fs import build_fs_handlers
 from ._handlers_theme import build_theme_handlers
 from ._persistence import (
     _atomic_write_json,
@@ -2363,159 +2364,6 @@ def build_settings_handlers(
             }
         )
 
-    async def _fs_custom_paths_call(service: str, data: dict[str, Any]) -> Any:
-        """Invoke a ha_mcp_tools component service for the custom-paths editor,
-        in any deployment mode (issue #1567).
-
-        Uses the live server's HA client in HTTP/add-on modes; in the stdio
-        sidecar (``server is None``) builds a transient ``HomeAssistantClient``
-        from the HA URL/token the sidecar inherits, and closes it afterward.
-        The caller wraps this in try/except so an unreachable HA / missing
-        token (e.g. OAuth mode, where ``server.client`` has no request-scoped
-        token) degrades to an "unavailable" envelope rather than a 500.
-        """
-        from ..tools.tools_filesystem import call_mcp_tools_service
-
-        own_client = None
-        try:
-            if server is not None:
-                client = server.client
-            else:
-                from ..client.rest_client import HomeAssistantClient
-
-                client = own_client = HomeAssistantClient()
-            return await call_mcp_tools_service(client, service, data)
-        finally:
-            if own_client is not None and hasattr(own_client, "close"):
-                with contextlib.suppress(Exception):
-                    await own_client.close()
-
-    async def _get_fs_custom_paths(_: Request) -> JSONResponse:
-        """Return the user-configured extra filesystem directories from the
-        ha_mcp_tools component, plus the non-overridable deny floor for the UI
-        blurb (issue #1567).
-
-        Always 200s with an ``available`` flag: when filesystem tools are off,
-        the component is missing/too old, or HA is unreachable, ``available``
-        is False with a human-readable ``reason`` so the UI can show a disabled
-        section instead of an error.
-        """
-        from ..tools.tools_filesystem import is_filesystem_tools_enabled
-        from ..tools.util_helpers import unwrap_service_response
-
-        def _unavailable(reason: str) -> JSONResponse:
-            return JSONResponse(
-                {
-                    "success": True,
-                    "available": False,
-                    "reason": reason,
-                    "paths": [],
-                    "deny_floor": [],
-                }
-            )
-
-        if not is_filesystem_tools_enabled():
-            return _unavailable(
-                "Filesystem tools are disabled. Enable them (beta) to "
-                "configure custom directories."
-            )
-        try:
-            result = await _fs_custom_paths_call("get_allowed_paths", {})
-        except Exception as exc:
-            logger.warning("fs-custom-paths GET could not reach ha_mcp_tools: %s", exc)
-            return _unavailable(f"Could not reach the ha_mcp_tools component: {exc}")
-
-        data = unwrap_service_response(result) if isinstance(result, dict) else {}
-        if not isinstance(data, dict) or not data.get("success", False):
-            reason = (
-                data.get("error") if isinstance(data, dict) else None
-            ) or "ha_mcp_tools returned an unexpected response."
-            return _unavailable(str(reason))
-        return JSONResponse(
-            {
-                "success": True,
-                "available": True,
-                "paths": data.get("paths", []),
-                "deny_floor": data.get("deny_floor", []),
-                "builtin_read_dirs": data.get("builtin_read_dirs", []),
-                "builtin_write_dirs": data.get("builtin_write_dirs", []),
-            }
-        )
-
-    async def _save_fs_custom_paths(request: Request) -> JSONResponse:
-        """Replace the user-configured extra filesystem directories via the
-        ha_mcp_tools component (issue #1567).
-
-        The component validates each entry and drops anything that hits the
-        deny floor or escapes the config dir; the dropped entries come back in
-        ``rejected``. ``restart_required`` is False — the component applies the
-        new allowlist live.
-        """
-        from ..tools.tools_filesystem import is_filesystem_tools_enabled
-        from ..tools.util_helpers import unwrap_service_response
-
-        if not is_filesystem_tools_enabled():
-            return JSONResponse(
-                create_error_response(
-                    ErrorCode.VALIDATION_INVALID_PARAMETER,
-                    "Filesystem tools are disabled; enable them (beta) before "
-                    "configuring custom directories.",
-                ),
-                status_code=409,
-            )
-        try:
-            body = await request.json()
-        except (ValueError, TypeError):
-            return JSONResponse(
-                create_error_response(
-                    ErrorCode.VALIDATION_INVALID_JSON,
-                    "Request body must be valid JSON.",
-                ),
-                status_code=400,
-            )
-        paths = body.get("paths") if isinstance(body, dict) else None
-        if paths is None:
-            paths = []
-        if not isinstance(paths, list) or not all(isinstance(p, str) for p in paths):
-            return JSONResponse(
-                create_error_response(
-                    ErrorCode.VALIDATION_INVALID_PARAMETER,
-                    "'paths' must be a list of directory strings.",
-                ),
-                status_code=400,
-            )
-        try:
-            result = await _fs_custom_paths_call("set_allowed_paths", {"paths": paths})
-        except Exception as exc:
-            logger.warning("fs-custom-paths POST could not reach ha_mcp_tools: %s", exc)
-            return JSONResponse(
-                create_error_response(
-                    ErrorCode.SERVICE_CALL_FAILED,
-                    f"Could not reach the ha_mcp_tools component: {exc}",
-                ),
-                status_code=502,
-            )
-
-        data = unwrap_service_response(result) if isinstance(result, dict) else {}
-        if not isinstance(data, dict) or not data.get("success", False):
-            reason = (
-                data.get("error") if isinstance(data, dict) else None
-            ) or "ha_mcp_tools rejected the update."
-            return JSONResponse(
-                create_error_response(ErrorCode.SERVICE_CALL_FAILED, str(reason)),
-                status_code=502,
-            )
-        return JSONResponse(
-            {
-                "success": True,
-                "applied": data.get("paths", []),
-                "paths": data.get("paths", []),
-                "rejected": data.get("rejected", []),
-                "mode": "component",
-                "restart_required": False,
-            }
-        )
-
     handlers: dict[str, Any] = {
         "root_page": _root_page,
         "settings_page": _settings_page,
@@ -2535,8 +2383,6 @@ def build_settings_handlers(
         "delete_backups_bulk": _delete_backups_bulk,
         "get_backup_config": _get_backup_config,
         "save_backup_config": _save_backup_config,
-        "get_fs_custom_paths": _get_fs_custom_paths,
-        "save_fs_custom_paths": _save_fs_custom_paths,
     }
 
     # Tool security policies. The main server attaches an
@@ -2563,6 +2409,7 @@ def build_settings_handlers(
 
     handlers.update(_build_visibility_handlers(data_dir=get_data_dir()))
     handlers.update(build_theme_handlers())
+    handlers.update(build_fs_handlers(server))
 
     return handlers
